refactor: contract / router / schema 분리 (#47)
