+++ src/schemas.ts
@@ -1,16 +1,13 @@
 import { z } from "zod";
 
-export const UserSchema = z.object({
-  user_id: z.string(),
-  password: z.string(),
-  lat: z.number(),
-  lon: z.number(),
-  is_recommend_lunch: z.boolean(),
-});
-
-export const UserSettingSchema = z.object({
-  user_id: z.string(),
-  lat: z.number(),
-  lon: z.number(),
-  is_recommend_lunch: z.boolean(),
+export const BaseHeadersSchema = z.object({
+  authorization: z.string(),
+  user: z
+    .object({
+      userId: z.string(),
+      lat: z.number(),
+      lon: z.number(),
+      isRecommendLunch: z.coerce.boolean(),
+    })
+    .optional(),
 });
